+++ plugins/libs/plugins-runtime/src/lib/api/index.ts
@@ -380,6 +380,11 @@ export function createApi(
       checkPermission('content:write');
       return plugin.context.createVariantFromComponents(shapes);
     },
+
+    waitForLayoutUpdate(timeout?: number): Promise<void> {
+      checkPermission('content:read');
+      return plugin.context.waitForLayoutUpdate(timeout);
+    },
   };
 
   return {